==> workflows/0. PDBP-clinical_qc_modelscompare.ipynb <==
 # Explore AMP-PD data available for the LID project

## Load libs, env variables, utils, and get path to data

# Use the os package to interact with the environment
import os

# Bring in Pandas for Dataframe functionality
import pandas as pd

# numpy for basics
import numpy as np

# To do tome regex ops
import re

# Use StringIO for working with file contents
from io import StringIO

# Enable IPython to display matplotlib graphs
import matplotlib.pyplot as plt
%matplotlib inline

# Enable interaction with the FireCloud API
from firecloud import api as fapi

# Import the iPython HTML rendering for displaying links to Google Cloud Console
from IPython.core.display import display, HTML

# Import urllib modules for building URLs to Google Cloud Console
import urllib.parse

# BigQuery for querying data
from google.cloud import bigquery

import sys

# pip install some python pkg to do cox proportional hazard analysis
! pip install lifelines
! pip show lifelines

%%capture 

sys.path.insert(0, '/home/jupyter/.local/lib/python3.7/site-packages')
from lifelines import CoxPHFitter

BILLING_PROJECT_ID = os.environ['GOOGLE_PROJECT']
WORKSPACE_NAMESPACE = os.environ['WORKSPACE_NAMESPACE']
WORKSPACE_NAME = os.environ['WORKSPACE_NAME']
WORKSPACE_BUCKET = os.environ['WORKSPACE_BUCKET']


WORKSPACE_ATTRIBUTES = fapi.get_workspace(WORKSPACE_NAMESPACE, WORKSPACE_NAME).json().get('workspace',{}).get('attributes',{})

GS_RELEASE_PATH = 'gs://amp-pd-data/releases/2021_v2-5release_0510'
GS_CLINICAL_RELEASE_PATH = f'{GS_RELEASE_PATH}/clinical/'

GS_WGS_RELEASE_PATH = 'gs://amp-pd-genomics/releases/2021_v2-5release_0510/wgs'
GS_WGS_RELEASE_PLINK_PATH = os.path.join(GS_WGS_RELEASE_PATH, 'plink')

BQ_RELEASE_DATASET = 'amp-pd-research.2021_v2-5release_0510'

# Utility routine for printing a shell command before executing it
def shell_do(command):
    print(f'Executing: {command}')
    !$command

def shell_return(command):
    print(f'Executing: {command}', file=sys.stderr)
    output = !$command
    return '\n'.join(output)


# Utility routine for display a message and a link
def display_html_link(description, link_text, url):
    html = f'''
    <p>
    </p>
    <p>
    {description}
    <a target=_blank href="{url}">{link_text}</a>.
    </p>
    '''

    display(HTML(html))
    
# Utility routines for reading files from Google Cloud Storage
def gcs_read_file(path):
    """Return the contents of a file in GCS"""
    contents = !gsutil -u {BILLING_PROJECT_ID} cat {path}
    return '\n'.join(contents)
    
def gcs_read_csv(path, sep=None):
    """Return a DataFrame from the contents of a delimited file in GCS"""
    return pd.read_csv(StringIO(gcs_read_file(path)), sep=sep, engine='python')

# Utility routine for displaying a message and link to Cloud Console
def link_to_cloud_console_gcs(description, link_text, gcs_path):
    url = '{}?{}'.format(
        os.path.join('https://console.cloud.google.com/storage/browser',
                     gcs_path.replace("gs://","")),
        urllib.parse.urlencode({'userProject': BILLING_PROJECT_ID}))

    display_html_link(description, link_text, url)
    
# Get the data from a query
def bq_query(query):
    print(f'Executing: {query}', file=sys.stderr)
    return pd.read_gbq(query, project_id=BILLING_PROJECT_ID, dialect='standard')

# Utility routine for displaying a message and link to Cloud Console
def link_to_cloud_console_gcs(description, link_text, gcs_path):
    url = '{}?{}'.format(
        os.path.join('https://console.cloud.google.com/storage/browser',
                     gcs_path.replace("gs://","")),
        urllib.parse.urlencode({'userProject': BILLING_PROJECT_ID}))

    display_html_link(description, link_text, url)
    
# Utility routine for displaying a message and link to Cloud Console
def link_to_cloud_console_bq(description, link_text, bq_dataset, bq_table=None):
    project, dataset = bq_dataset.split('.', 1)
    if bq_table:
        page_params = {'page': 'table', 'p': project, 'd': dataset, 't': bq_table}
    else:
        page_params = {'page': 'dataset', 'p': project, 'd': dataset}
    
    url = '{}?{}'.format(
        'https://console.cloud.google.com/bigquery',
        urllib.parse.urlencode(page_params))

    display_html_link(description, link_text, url)    

# Utility routine for printing a query before executing it
def bq_query(query):
    """Return the contents of a query against BigQuery"""
    return pd.read_gbq(
        query,
        project_id=BILLING_PROJECT_ID,
        dialect='standard')

shell_do(f'gsutil -u {BILLING_PROJECT_ID} ls {GS_RELEASE_PATH}')
#shell_do(f'gsutil -u {BILLING_PROJECT_ID} ls {GS_CLINICAL_RELEASE_PATH}')

shell_do(f'gsutil -mu {BILLING_PROJECT_ID} ls {WORKSPACE_BUCKET}')

## Get and explore AMP-PD PD data

pd_case_control_df = gcs_read_csv(os.path.join(GS_RELEASE_PATH, 'amp_pd_case_control.csv'))
pd_case_control_df.info()
#pd_case_control_df.head()

# Subset and keep only latest diagnosis 
pd_case_control_df_latest = pd_case_control_df[['participant_id', 'diagnosis_latest', 'case_control_other_latest']]
pd_case_control_df_latest.columns = ['ID', 'LATEST_DX', 'CASE_CONTROL']

#print(pd_case_control_df_noPPMI_latest.value_counts())
print(pd_case_control_df_latest['LATEST_DX'].value_counts())

# Grab the enrollment info so we know how each patient was recruited 
enrollment_df = gcs_read_csv(os.path.join(GS_CLINICAL_RELEASE_PATH, 'Enrollment.csv'))
enrollment_subset_df = enrollment_df[['participant_id', 'study_arm']].copy()
enrollment_subset_df.columns = ['ID', 'ENROLL_STUDY_ARM']

print(enrollment_subset_df['ENROLL_STUDY_ARM'].value_counts(dropna=False))

demo_enrollment_df = pd_case_control_df_latest.merge(enrollment_subset_df, on="ID", how="left")
demo_enrollment_df.info()
#demo_enrollment_df.head()

# Coding for PD_PHENO in AMP data
    # 1=Healthy control with no neurological disorders
    # 2=Positive PD case not in a genetic enrichment study, prodromal, or a SWEDD
conditions = [
    (demo_enrollment_df['ENROLL_STUDY_ARM'] == "Healthy Control") & (demo_enrollment_df['LATEST_DX'] == "No PD Nor Other Neurological Disorder"),
    (demo_enrollment_df['ENROLL_STUDY_ARM'] == "Healthy Control") & (demo_enrollment_df['LATEST_DX'] == "Idiopathic PD"),
    (demo_enrollment_df['ENROLL_STUDY_ARM'] == "PD") & (demo_enrollment_df['LATEST_DX'] == "Parkinson's Disease"),
    (demo_enrollment_df['ENROLL_STUDY_ARM'] == "PD") & (demo_enrollment_df['LATEST_DX'] == "Idiopathic PD")
]
choices = [1,3,2,2]

demo_enrollment_df['PD_PHENO'] = np.select(conditions, choices, default=-9)
print(demo_enrollment_df['PD_PHENO'].value_counts(dropna=False))
#demo_enrollment_df.loc[demo_enrollment_df['PD_PHENO'] == 3]

## Add study as a column 
# Add cohort column (Numpy's NAN for anything missing - though nothing should be missing)
demo_enrollment_df['COHORT'] = np.where(demo_enrollment_df.ID.str.contains("LB-"), "LBD",
                               np.where(demo_enrollment_df.ID.str.contains("PD-"), "PDBP",
                               np.where(demo_enrollment_df.ID.str.contains("PP-"), "PPMI",
                               np.where(demo_enrollment_df.ID.str.contains("HB-"), "HBS",
                               np.where(demo_enrollment_df.ID.str.contains("LC-"), "LCC",
                               np.where(demo_enrollment_df.ID.str.contains("BF-"), "BIOFIND",
                               np.where(demo_enrollment_df.ID.str.contains("SU-"), "SURE-PD3",
                               np.where(demo_enrollment_df.ID.str.contains("SY-"), "STEADY-PD3", np.nan))))))))

print(demo_enrollment_df['COHORT'].value_counts(dropna=True))
demo_enrollment_df.info()

demo_enrollment_df_PD = demo_enrollment_df[demo_enrollment_df.PD_PHENO.eq(2)]
len(demo_enrollment_df_PD.ID.unique())
demo_enrollment_df_PD.info()
#demo_enrollment_df_PD.head()

## Load and explore MDSUPDRSIV data

MDSUPDRSIV_dict = gcs_read_csv(os.path.join(GS_CLINICAL_RELEASE_PATH, 'MDS_UPDRS_Part_IV_dictionary.csv'))
MDSUPDRSIV_dict.info()
#with pd.option_context('display.max_rows', None, 'display.max_columns', None):  # more options can be specified also
#    display(MDSUPDRSIV_dict)

### Potential data that can be used in LID study

MDSUPDRSIV = gcs_read_csv(os.path.join(GS_CLINICAL_RELEASE_PATH, 'MDS_UPDRS_Part_IV.csv'))
MDSUPDRSIV = MDSUPDRSIV[["participant_id", "GUID", "visit_name", "visit_month", "code_upd2401_time_spent_with_dyskinesias"]]
MDSUPDRSIV.columns = ['ID', 'GUID', 'visit_name', 'visit_month', 'MDS_UPDRS_IV_1']

## Add study as a column 
# Add cohort column (Numpy's NAN for anything missing - though nothing should be missing)
MDSUPDRSIV['COHORT'] = np.where(MDSUPDRSIV.ID.str.contains("LB-"), "LBD",
                               np.where(MDSUPDRSIV.ID.str.contains("PD-"), "PDBP",
                               np.where(MDSUPDRSIV.ID.str.contains("PP-"), "PPMI",
                               np.where(MDSUPDRSIV.ID.str.contains("HB-"), "HBS",
                               np.where(MDSUPDRSIV.ID.str.contains("LC-"), "LCC",
                               np.where(MDSUPDRSIV.ID.str.contains("BF-"), "BIOFIND",
                               np.where(MDSUPDRSIV.ID.str.contains("SU-"), "SURE-PD3",
                               np.where(MDSUPDRSIV.ID.str.contains("SY-"), "STEADY-PD3", np.nan))))))))

print(MDSUPDRSIV[['COHORT', 'visit_month']].value_counts(dropna=True).sort_index())
MDSUPDRSIV.info()




    **CONCLUSSIONS**  
PDBP and SURE-PD3 seems good candidates to be used in the LID analysis

demo_enrollment_df_PD_filtered = demo_enrollment_df_PD[demo_enrollment_df_PD['ID'].str.contains("PD-|SU-")]
demo_enrollment_df_PD_filtered = demo_enrollment_df_PD_filtered.drop('COHORT', axis=1)
demo_enrollment_df_PD_filtered

# We inner join the demo_PD file. Then, we come up with all PD patients from 
MDSUPDRSIV_DEMO = demo_enrollment_df_PD_filtered.drop_duplicates().merge(MDSUPDRSIV, 
                                                                         left_on=['ID'], right_on=['ID'],
                                                                         how='inner')

MDSUPDRSIV_DEMO.info()

### Explore NAs in the outcome. Look at NAs stratified by cohort

MDSUPDRSIV_DEMO.MDS_UPDRS_IV_1.isnull().groupby([MDSUPDRSIV_DEMO['COHORT'], MDSUPDRSIV_DEMO['visit_month']]).sum().astype(int)

#Look at the number of ppl meeting the outcome at each visit
MDSUPDRSIV_DEMO.MDS_UPDRS_IV_1.ge(2).groupby([MDSUPDRSIV_DEMO['COHORT'], MDSUPDRSIV_DEMO['visit_month']]).sum().astype(int)


### Conclussion

**CONCLUSSION**  
I think the NA pattern we see in SURE-PD3 is because MDS-UPDRS IV assessment is not taken until L-dopa treatment starts


**Make decision regarding using SURE-PD3 data or not**

DECISION: There are almost no people developing dyskinesias in SURE-PD. I will exclude this, and work only with PDBP

## PDBP QC

MDSUPDRSIV_DEMO_PDBPD = MDSUPDRSIV[MDSUPDRSIV['ID'].str.contains("PD-")]
MDSUPDRSIV_DEMO_PDBPD.head()

MDSUPDRSIV_DEMO_PDBPD.ID.nunique()


### Remove ppl that have non PD diagnosis

# We inner join the demo_PD file. Then, we come up with all PD patients from 
MDSUPDRSIV_DEMO_PDBPD = demo_enrollment_df_PD_filtered.drop_duplicates().merge(MDSUPDRSIV_DEMO_PDBPD, 
                                                                          left_on=['ID'], right_on=['ID'],
                                                                          how='inner')
MDSUPDRSIV_DEMO_PDBPD.head()
MDSUPDRSIV_DEMO_PDBPD.ID.nunique()

### Remove people that have never had levodopa during the length of the study

medical_hist = gcs_read_csv(os.path.join(GS_CLINICAL_RELEASE_PATH, 'PD_Medical_History.csv'))
medical_hist.info()
#medical_hist.head()
print("Levodopa info values", medical_hist.on_levodopa.unique())


medical_hist_PDBP = medical_hist[medical_hist["participant_id"].isin(MDSUPDRSIV_DEMO_PDBPD.ID.unique())]
medical_hist_PDBP.shape

#Checking we have the expacted number of PD patients

print("\nNumber of PD patients with medical history" ,len(medical_hist_PDBP.participant_id.unique()))

medical_hist_PDBP_LDOPA = medical_hist_PDBP.loc[(medical_hist_PDBP['on_levodopa']=='Yes')]
print("\nNumber of PD patients that have taken Leovodpa at least once" ,len(medical_hist_PDBP_LDOPA.participant_id.unique()))
medical_hist_PDBP_LDOPA.head()

# We save these patients IDs
PD_IDs_QC = set(medical_hist_PDBP_LDOPA.participant_id.unique())

medical_hist_PDBP_LDOPA.on_levodopa.unique()

### Identify ppl with disease duration from BL to diagnosis > 10 years

Rationale of doing this is that the period of uncertainty is too wide.  
These people could have developed LID any time within those 6 years  
Then, treatment amendment according to LID could prevent us from detecting potential LID devevelopers

#medical_hist_PDBP_LDOPA_BL = medical_hist_PDBP_LDOPA[
#    medical_hist_PDBP_LDOPA[["participant_id", "age_at_diagnosis", "visit_name"]]["visit_name"] == "M0"]
#medical_hist_PDBP_LDOPA_BL
AAD = medical_hist_PDBP_LDOPA[["participant_id", "visit_name", "age_at_diagnosis"]][medical_hist_PDBP_LDOPA['age_at_diagnosis'].notnull()]
# Get all non na entries a
AAD_QC = AAD.drop_duplicates(subset = "participant_id")


# Filter the AAD_QC by the total number of samples with 

# We are loosing some samples that do not have AAD
print("Number of samples with AAD data:",AAD_QC.shape[0])
#print(AAD_QC.head())



#### Keep LID developers missing AAD

AAD = medical_hist_PDBP_LDOPA[["participant_id", "visit_name", "age_at_diagnosis"]][medical_hist_PDBP_LDOPA['age_at_diagnosis'].notnull()]
# Get all non na entries a
AAD_QC = AAD.drop_duplicates(subset = "participant_id")

# We are loosing some samples that do not have AAD
print("Number of samples with AAD data:",AAD_QC.shape[0])

# Keep people missing AAD and developing LID. 
# If they are LID developers I will consier includind them
AAD_MISSING = set(PD_IDs_QC).difference(set(list(AAD_QC.participant_id)))

# Check if AAD_MISSING ppl develop dyskinesias
MDSUPDRSIV_AAD_MISSING = MDSUPDRSIV_DEMO_PDBPD[MDSUPDRSIV_DEMO_PDBPD["ID"].isin(AAD_MISSING)]
MDSUPDRSIV_AAD_MISSING.MDS_UPDRS_IV_1.ge(2).groupby([MDSUPDRSIV_AAD_MISSING['COHORT'], MDSUPDRSIV_AAD_MISSING['visit_month']]).sum().astype(int)

# Get the IIDs of the three people that develop LID and that are missing AAD.
LID_AADMISSING = set(MDSUPDRSIV_AAD_MISSING[MDSUPDRSIV_AAD_MISSING.MDS_UPDRS_IV_1.ge(2)]["ID"])


# We exclude the LID_AADMISSING ppl from the AAD_MISSING set
AAD_MISSING_EXCLUDE = AAD_MISSING - LID_AADMISSING
[PD_IDs_QC.remove(i) for i in AAD_MISSING_EXCLUDE]
print("\nNumber of PD patients that have AAD including three imputed samples" ,len(PD_IDs_QC))

# Then, we include these 3 ppl in AAOD_QC
#AAD = medical_hist_PDBP_LDOPA[["participant_id", "visit_name", "age_at_diagnosis"]][medical_hist_PDBP_LDOPA["participant_id"].isin(PD_IDs_QC)]
#AAD_QC = AAD.drop_duplicates(subset = "participant_id")
#print("Number of samples with AAD data including 3 LID developers:",AAD_QC.shape[0])

len(PD_IDs_QC)

# get age at BL from demographics df and AAD from medical history. Then merge
demographics = gcs_read_csv(os.path.join(GS_CLINICAL_RELEASE_PATH, 'Demographics.csv'))
demographics.info()

demo_AAD = pd.merge(demographics[demographics["visit_name"] == "M0"],
                    AAD_QC[["participant_id", "age_at_diagnosis"]],
                    on = "participant_id",
                    how = "right")

# Calculate disease duration from diagnosis
demo_AAD["DIS_DUR"] = demo_AAD.age_at_baseline - demo_AAD.age_at_diagnosis

demo_AAD_DISDURLONG = list(demo_AAD[demo_AAD.DIS_DUR.ge(9)].participant_id)
len(demo_AAD_DISDURLONG)

# There are 70 ppl with long disease duration.
# We will perform an initial run with and without including these

MDSUPDRSIV_DEMO_PDBPD = MDSUPDRSIV_DEMO_PDBPD[~MDSUPDRSIV_DEMO_PDBPD["ID"].isin(demo_AAD_DISDURLONG)]


MDSUPDRSIV_DEMO_PDBPD.ID.unique().size

### Remove patients that only have CS data

# Look at the number of entries at each visit
#MDSUPDRSIV_DEMO_PDBPD.groupby(['visit_name'])['ID'].count()

ID_COUNTS = MDSUPDRSIV_DEMO_PDBPD.groupby(['ID'])['visit_name'].size().reset_index(name='counts')

ID_EXCLUDE = list(ID_COUNTS[ID_COUNTS.counts.eq(1)].ID)

MDSUPDRSIV_DEMO_PDBPD_LG=MDSUPDRSIV_DEMO_PDBPD[~MDSUPDRSIV_DEMO_PDBPD["ID"].isin(ID_EXCLUDE)]

MDSUPDRSIV_DEMO_PDBPD_LG.shape

print("\n\n",len(MDSUPDRSIV_DEMO_PDBPD_LG.ID.unique()),"Patients with longitudinal data available")
#MDSUPDRSIV_DEMO_PDBPD_LG.head()

MDSUPDRSIV_DEMO_PDBPD_LG.groupby(['visit_name'])['visit_name'].count()

### Remove left censored people

In 1.3.2 we could see how there is a total of 45 PD patients that already had dyskinesias at BL.
Remove these.


# Get baseline data and retrieve the IDs of ppl with MDSUDPRS_IV ge(2)

CENSORED_PATIENTS = list(MDSUPDRSIV_DEMO_PDBPD_LG.loc[(MDSUPDRSIV_DEMO_PDBPD_LG['visit_name']=="M0") & 
                             (MDSUPDRSIV_DEMO_PDBPD_LG['MDS_UPDRS_IV_1']>=2)].ID)

MDSUPDRSIV_DEMO_PDBPD_LG_FILT=MDSUPDRSIV_DEMO_PDBPD_LG[~MDSUPDRSIV_DEMO_PDBPD_LG["ID"].isin(CENSORED_PATIENTS)]

MDSUPDRSIV_DEMO_PDBPD_LG_FILT.shape
print("\n\n",len(MDSUPDRSIV_DEMO_PDBPD_LG_FILT.ID.unique()),"No Left censored patients")

### Time to LID outcomes and getting all covariates

#### Get COVARIATES

# We will use the list of PDBP patients saved in 1.4.3 section
# Then, get demographics
PD_IDs_QC = list(MDSUPDRSIV_DEMO_PDBPD_LG_FILT.ID.unique())
PDBP_QC = pd.merge(pd.DataFrame(PD_IDs_QC, columns =['ID']),
                   demographics[["participant_id", "age_at_baseline", "sex"]],
                   left_on = "ID", right_on = "participant_id",
                   how = "left")

# Getting SEX
PDBP_QC['sex'] = np.where(PDBP_QC['sex'] == "Female", 2, 1)


# Get the AAD info
PDBP_QC = pd.merge(PDBP_QC,
                   AAD_QC.drop("visit_name", axis = 1),
                   on = "participant_id",
                   how = "left")
# Check there are 3 ppl missing AAD
#print("Number NAs", PDBP_QC['age_at_diagnosis'].isna().sum())
# Impute the AAD for those three ppl from the AAB
print("Check the total of ppl with AAB NAs: ", PDBP_QC.age_at_baseline.isna().sum())
PDBP_QC['AAD_MISSING'] = PDBP_QC['age_at_diagnosis'].isna().astype(int)
PDBP_QC["AAD_IMPUTED"] = PDBP_QC.age_at_diagnosis.mask(PDBP_QC.age_at_diagnosis.isna(),
                                                      PDBP_QC.age_at_baseline - 
                                                       (PDBP_QC.age_at_baseline - PDBP_QC.age_at_diagnosis.mean()))

# Get a tag to identify ppl with long disease duration

# Recalculate DISEASE DURATION after including 3 LID ppl with AAD imputed
# Calculate disease duration from diagnosis
PDBP_QC["DIS_DUR"] = PDBP_QC.age_at_baseline - PDBP_QC.AAD_IMPUTED
DISDURLONG = list(PDBP_QC[PDBP_QC.DIS_DUR.ge(9)].participant_id)
print("PPL WITH LONG DISEASE DURATION: ", len(DISDURLONG))
PDBP_QC.head()
# Add tag for the 70 ppl with long disease duration. I will exclude them on sensitivity analyses
PDBP_QC['LONG_DISDUR'] = PDBP_QC['participant_id'].isin(DISDURLONG).astype(int)


PDBP_QC["DIS_DUR"] =  PDBP_QC.age_at_baseline - PDBP_QC.AAD_IMPUTED
# Check entried with a negative disease duration
PDBP_QC[PDBP_QC.DIS_DUR < 0]
# Exclude patient with DIS_DUR < -10 Years. Possible error in the recoding for that entry
PDBP_QC = PDBP_QC[~PDBP_QC.ID.str.contains('PD-PDAW438DAN')]
# In addition, set the DISEASE DURATION TO 0 for PPL with a disease duration of -1 or -2
PDBP_QC["DIS_DUR"] = [0 if dur < 0 else dur for dur in PDBP_QC.DIS_DUR]
PDBP_QC.info()
print("\nChecking disease duration is okey \n")
print(PDBP_QC[PDBP_QC.DIS_DUR < 0])






# Add the MDS-UPDRSIIItotal score at BL
#demographics = gcs_read_csv(os.path.join(GS_CLINICAL_RELEASE_PATH, 'MDS_UPDRS_Part_III.csv'))
MDSUPDRSIII = gcs_read_csv(os.path.join(GS_CLINICAL_RELEASE_PATH, 'MDS_UPDRS_Part_III.csv'))
MDSUPDRSIII_BL = MDSUPDRSIII[MDSUPDRSIII.visit_name == "M0"]
PDBP_QC = pd.merge(PDBP_QC,
                   MDSUPDRSIII_BL[["participant_id", "mds_updrs_part_iii_summary_score"]],
                   on = "participant_id",
                   how = "left")
print("Check the number of MDSPUPDRSIII_BL NAs: ", PDBP_QC['mds_updrs_part_iii_summary_score'].isna().sum())

#### Get OUTCOME

# GETTING THE TIME EVENT OUTCOME
MDSUPDRSIV_OUTCOMES = MDSUPDRSIV_DEMO_PDBPD_LG_FILT[MDSUPDRSIV_DEMO_PDBPD_LG_FILT["ID"].isin(PD_IDs_QC)]


# Get all entries with MDS-UPDRS-IV-1 >= 2
# Group by ID
# Get the first entry (ie ming visit_month value)
MDSUPDRSIV_LID = MDSUPDRSIV_OUTCOMES[MDSUPDRSIV_OUTCOMES.MDS_UPDRS_IV_1.ge(2)]
idx = MDSUPDRSIV_LID.groupby(['ID'])['visit_month'].transform(min) == MDSUPDRSIV_LID['visit_month']
MDSUPDRSIV_LID = MDSUPDRSIV_LID[idx]
# To define the event midpoint, we just substract the visit months ( assessments were performed eveyr 3 months)
MDSUPDRSIV_LID["TIMEMIDPOINT_YEARS"] = (MDSUPDRSIV_LID.visit_month - 3 ) / 12
MDSUPDRSIV_LID["EVENT_DYSK"] = 1

# Get another subset filtering out previous samples
# Group by ID
# Get the last entry (ie ming visit_month value)
MDSUPDRSIV_NOLID = MDSUPDRSIV_OUTCOMES[~MDSUPDRSIV_OUTCOMES.ID.isin(MDSUPDRSIV_LID.ID)]


idx = MDSUPDRSIV_NOLID.groupby(['ID'])['visit_month'].transform(max) == MDSUPDRSIV_NOLID['visit_month']
MDSUPDRSIV_NOLID = MDSUPDRSIV_NOLID[idx]
# To define the event midpoint, we just substract the visit months ( assessments were performed eveyr 3 months)
MDSUPDRSIV_NOLID["TIMEMIDPOINT_YEARS"] = [month / 12 if month == 60 else (month + 3) / 12  
                                               for month in MDSUPDRSIV_NOLID.visit_month]

MDSUPDRSIV_NOLID["EVENT_DYSK"] = 0
print("\nNumber of patients with LID",len(MDSUPDRSIV_LID.ID.unique()))
print("\nNumber of patients without LID",len(MDSUPDRSIV_NOLID.ID.unique()))

MDSUPDRSIV_TIMEPOINT = pd.concat([MDSUPDRSIV_LID[["ID","visit_month", "TIMEMIDPOINT_YEARS", "EVENT_DYSK"]], 
                                  MDSUPDRSIV_NOLID[["ID", "visit_month", "TIMEMIDPOINT_YEARS", "EVENT_DYSK"]]])
MDSUPDRSIV_TIMEPOINT.info()

#with pd.option_context('display.max_rows', None, 'display.max_columns', None):  # more options can be specified also
#    tmp = MDSUPDRSIV_LID.sort_values(["ID", "visit_month"])
#    display(tmp)
#    display(MDSUPDRSIV_LID)

#### Merge the outcome with covariates

PDBP_QC = pd.merge(PDBP_QC,
                   MDSUPDRSIV_TIMEPOINT,
                   on = "ID",
                   how = "left")


PDBP_QC = PDBP_QC.rename(columns={"mds_updrs_part_iii_summary_score" : "MDSUPDRSIII_BL", "sex" : "SEX"}, errors="raise")

PDBP_QC.drop(["participant_id"], axis=1, inplace=True)
PDBP_QC.info()
#PDBP_QC.head()

#### Get the Ldopa data at the time the event is met
We are going to load the Levodopa only LEDD data Tarek generated some time ago and we will append it to the generated DF

shell_do(f'gsutil ls {WORKSPACE_BUCKET}')
ledd_levo = gcs_read_csv(os.path.join(WORKSPACE_BUCKET, 'clinical_data/PDBP_derived_ledd.csv'))
ledd_levo["ID"] = ("PD-" + ledd_levo.SGUID.map(lambda x: re.sub(r"^\d+\.", "", x))).astype(str)
#ledd_levo

ledd_levo_bl = ledd_levo[ledd_levo.VisitTypPDBP == "Baseline"]
ledd_levo_bl_qc = ledd_levo_bl[ledd_levo_bl["ID"].isin(PDBP_QC.ID.unique())]
ledd_levo_bl_qc
ledd_levo_bl_qc.describe(include=np.number) 

shell_do(f'gsutil ls {WORKSPACE_BUCKET}')
ledd_levo = gcs_read_csv(os.path.join(WORKSPACE_BUCKET, 'clinical_data/PDBP_derived_ledd.csv'))
ledd_levo["ID"] = ("PD-" + ledd_levo.SGUID.map(lambda x: re.sub(r"^\d+\.", "", x))).astype(str)

#df['col'] = 'str' + df['col'].astype(str)

ledd_levo["visit_month"] = ledd_levo.VisitTypPDBP.map(lambda x: re.sub(r"months$", "", x))
ledd_levo["visit_month"] = pd.to_numeric(np.where(ledd_levo.visit_month == "Baseline", "0", ledd_levo.visit_month), errors='raise')

ledd_levo.info()
PDBP_QC = pd.merge(PDBP_QC,
                   ledd_levo[["ID", "visit_month", "LEDD", "DailyLevo"]], 
                   on = ["ID", "visit_month"],
                   #on = "ID",
                   how = "left")

PDBP_QC.info()

**NOTE TO MYSELF**
Levodop data is unsuitable.  
We are missing data for half of the patients in the QCed data

### PDBP summary statistics
Get metrics for the whole cohort and stratified by event_dyskinesias

PDBP_QC.LONG_DISDUR.unique()

PDBP_QC.info()

# Funny casualty is that the numbers below match. I did check this is not coming from an error during processing
print(PDBP_QC["EVENT_DYSK"].value_counts())
print(PDBP_QC["LONG_DISDUR"].value_counts())
# Check the outcome distribution

PDBP_QC.describe(include=np.number)  

# Checking number of missingness
PDBP_QC.isna().sum()

#### Observe the sumary statistitc stratified by the disease duration groups

with pd.option_context('display.max_rows', None, 'display.max_columns', None):  # more options can be specified also
    print(PDBP_QC.groupby('LONG_DISDUR').describe(include=np.number))   

### COX PH models fitting
Here, we want to have a look at th models and see whether the COXZP is met and whether the covariates are significantly associated with the time to event outcome

PDBP_QC_SEXAAD = PDBP_QC[["TIMEMIDPOINT_YEARS", "EVENT_DYSK", "AAD_IMPUTED", "SEX"]]


cph = CoxPHFitter()
cph.fit(PDBP_QC_SEXAAD, "TIMEMIDPOINT_YEARS", event_col = "EVENT_DYSK")
cph.print_summary()
cph.check_assumptions(PDBP_QC_SEXAAD, show_plots=True)

PDBP_QC_SEXAAD_DISDUR = PDBP_QC[["TIMEMIDPOINT_YEARS", "EVENT_DYSK", "AAD_IMPUTED", "SEX", "DIS_DUR"]]


cph = CoxPHFitter()
cph.fit(PDBP_QC_SEXAAD_DISDUR, "TIMEMIDPOINT_YEARS", event_col = "EVENT_DYSK")
cph.print_summary()

cph.check_assumptions(PDBP_QC_SEXAAD_DISDUR, show_plots=True)

PDBP_QC_SEXAAD_DISDUR_MDSUPDRSIII = PDBP_QC[["TIMEMIDPOINT_YEARS", "EVENT_DYSK", "AAD_IMPUTED", "SEX", "DIS_DUR",
                                "MDSUPDRSIII_BL"]]

cph = CoxPHFitter()
cph.fit(PDBP_QC_SEXAAD_DISDUR_MDSUPDRSIII, "TIMEMIDPOINT_YEARS", event_col = "EVENT_DYSK")
cph.print_summary()

cph.check_assumptions(PDBP_QC_SEXAAD_DISDUR_MDSUPDRSIII, show_plots=True)

PDBP_QC_SEXAAD_DISDUR_MDSUPDRSIII_LDOPA = PDBP_QC[["TIMEMIDPOINT_YEARS", "EVENT_DYSK", "AAD_IMPUTED", "SEX", "DIS_DUR",
                                "MDSUPDRSIII_BL", "DailyLevo"]]


PDBP_QC_SEXAAD_DISDUR_MDSUPDRSIII_LDOPA = PDBP_QC_SEXAAD_DISDUR_MDSUPDRSIII_LDOPA[PDBP_QC_SEXAAD_DISDUR_MDSUPDRSIII_LDOPA['DailyLevo'].notnull()]

cph = CoxPHFitter()
cph.fit(PDBP_QC_SEXAAD_DISDUR_MDSUPDRSIII_LDOPA, "TIMEMIDPOINT_YEARS", event_col = "EVENT_DYSK")
cph.print_summary()

cph.check_assumptions(PDBP_QC_SEXAAD_DISDUR_MDSUPDRSIII_LDOPA, show_plots=True)

### Sensitivity analysis
The idea is to exclude people with long disease duration and observe the results

PDBP_QC_NODISDUR = PDBP_QC[PDBP_QC.LONG_DISDUR.eq(0)]
PDBP_QC_NODISDUR.info()

print(PDBP_QC_NODISDUR["EVENT_DYSK"].value_counts())
print(PDBP_QC_NODISDUR["LONG_DISDUR"].value_counts())
# Check the outcome distribution

PDBP_QC_SEXAAD_DISDUR_MDSUPDRSIII = PDBP_QC_NODISDUR[["TIMEMIDPOINT_YEARS", "EVENT_DYSK", "AAD_IMPUTED", "SEX", "DIS_DUR",
                                "MDSUPDRSIII_BL"]]

cph = CoxPHFitter()
cph.fit(PDBP_QC_SEXAAD_DISDUR_MDSUPDRSIII, "TIMEMIDPOINT_YEARS", event_col = "EVENT_DYSK")
cph.print_summary()

cph.check_assumptions(PDBP_QC_SEXAAD_DISDUR_MDSUPDRSIII, show_plots=True)

PDBP_QC_SEXAAD_DISDUR_MDSUPDRSIII_LDOPA = PDBP_QC_NODISDUR[["TIMEMIDPOINT_YEARS", "EVENT_DYSK", "AAD_IMPUTED", "SEX", "DIS_DUR",
                                "MDSUPDRSIII_BL", "DailyLevo"]]


PDBP_QC_SEXAAD_DISDUR_MDSUPDRSIII_LDOPA = PDBP_QC_SEXAAD_DISDUR_MDSUPDRSIII_LDOPA[PDBP_QC_SEXAAD_DISDUR_MDSUPDRSIII_LDOPA['DailyLevo'].notnull()]


cph = CoxPHFitter()
cph.fit(PDBP_QC_SEXAAD_DISDUR_MDSUPDRSIII_LDOPA, "TIMEMIDPOINT_YEARS", event_col = "EVENT_DYSK")
cph.print_summary()

cph.check_assumptions(PDBP_QC_SEXAAD_DISDUR_MDSUPDRSIII_LDOPA, show_plots=True)

## SAVE PDBPD after QC


shell_do(f'gsutil ls {WORKSPACE_BUCKET}')
ledd_levo = gcs_read_csv(os.path.join(WORKSPACE_BUCKET, 'clinical_data/PDBP_derived_ledd.csv'))

# Save our merged dataframe out to a .csv and .txt file
csv_output = "PDBP_LID.csv"
txt_output = "PDBP_LID.txt"
PDBP_QC.to_csv(csv_output, index=False, sep=",")
PDBP_QC.to_csv(txt_output, index=False, sep="\t")


shell_do(f'gsutil ls {WORKSPACE_BUCKET}')

# Save this file to a new directory in your cloud storage 
shell_do(f'gsutil -u {BILLING_PROJECT_ID} -m cp {csv_output} {WORKSPACE_BUCKET}/clinical_data/')
shell_do(f'gsutil -u {BILLING_PROJECT_ID} -m cp {txt_output} {WORKSPACE_BUCKET}/clinical_data/')
