==> workflows/6. forest_plots.Rmd <==
---
title: "ForestPlots"
author: "AMCalejandro"
date: "27/09/2022"
output: html_document
---


# load data and forest plot custom function
```{r}
library(forestplot)
library(dplyr)
library(tidyverse)

source(here::here("utils/utils.R"))
```


# Get the per cohorts SNPs metrics to generate the forestplots

```{bash}
grep ".txt" META_SCRIPT_STDERR.txt  |  
  awk '{print $2}' | 
  xargs grep -E "4:32435284|16:17044975|1:53778300" >> percohort_metrics.txt
```



# Get forestplot for lead SNP in chromosome 1

```{r}
# rs72673189
SNP_cohort = tibble(
  mean = c(1.02, 1.38, 0.75, 0.56),
  lower= c(1.02 - (1.96*0.27), 1.38 - (1.96*0.34),
           0.75 - (1.96*0.40), 0.56 - (1.96*0.35)),
  upper= c(1.02 + (1.96*0.27), 1.38 + (1.96*0.34),
           0.75 + (1.96*0.40), 0.56 + (1.96*0.35)),
  study = c("TPD", "OPDC","PPMI", "PDBP"),
  effect = c("1.02", "1.38", "0.75", "0.56"),
  ci = c("[0.49, 1.55]", "[0.71, 2.05]", "[0, 1.53]", "[-0.13, 1.25]")
  )

summary = tibble(mean = 1.01,
                 lower = 1.01 - (1.96*0.17),
                 upper = 1.01 + (1.96*0.17),
                 study = "Summary",
                 effect = "1.01",
                 ci =  "[0.68 ,1.34]",
                 summary = TRUE)
header = tibble(study = "Study",
                effect = "β",
                ci = "CI",
                summary = TRUE)
empty_row <- tibble(mean = NA_real_)
SNPdf <- bind_rows(header,
                   SNP_cohort,
                   empty_row,
                   summary)


a = myforestplot(df = SNPdf, 
                 snpname = "rs72673189", 
                  n_studies = 4)

pdf("../FIGURES/rs72673189_forestPlot.pdf", width = 9)
a
dev.off()
jpeg("../FIGURES/rs72673189_forestPlot.jpg",
    width = 7.5, height = 6.5, units = "in", res = 500)
a 
dev.off()

```


# Get forestplot for lead SNP in chromosome 4

```{r}

# rs189093213
SNP_cohort = tibble(
  mean = c(0.95, 1.07, 1.73, 0.10, 1.87),
  lower= c(0.95 - (1.96*0.26), 1.07 - (1.96*0.40),
           1.73 - (1.96*0.40), 0.10 - (1.96*0.76),
           1.87 - (1.96*0.90)),
  upper= c(0.95 + (1.96*0.26), 1.07 + (1.96*0.40),
           1.73 + (1.96*0.40), 0.10 + (1.96*0.76),
           1.87 + (1.96*0.90)),
  study = c("TPD", "OPDC","PPMI", "PDBP", "PD-STAT"),
  effect = c("0.95", "1.07", "1.73", "0.10", "1.87"),
  ci = c("[0.44, 1.46]", "[0.29, 1.85]", "[0.95, 2.51]", 
         "[-1.3, 1.59]", "[0.1, 3.63]")
  )


summary = tibble(mean = 1.13,
                 lower = 1.13 - (1.96*0.18),
                 upper = 1.13 + (1.96*0.18),
                 study = "Summary",
                 effect = "1.13",
                 ci =  "[0.68 ,1.34]",
                 summary = TRUE)
header = tibble(study = "Study",
                effect = "β",
                ci = "CI",
                summary = TRUE)
empty_row <- tibble(mean = NA_real_)
SNPdf <- bind_rows(header,
                   SNP_cohort,
                   empty_row,
                   summary)

a = myforestplot(df = SNPdf, 
                 snpname = "rs189093213", 
                  n_studies = 5)


pdf("../FIGURES/rs189093213_forestPlot.pdf", width = 9)
a
dev.off()

jpeg("../FIGURES/rs189093213_forestPlot.jpg",
    width = 7.5, height = 6.5, units = "in", res = 500)
a 
dev.off()

```


# Get forestplot for lead SNP in chromosome 16

```{r}
# rs180924818
SNP_cohort = tibble(
  mean = c(1.08, 1.01, 1.49),
  lower= c(1.08 - (1.96*0.26), 1.01 - (1.96*0.40),
           1.49 - (1.96*0.40)),
  upper= c(1.08 + (1.96*0.26), 1.01 + (1.96*0.40),
           1.49 + (1.96*0.40)),
  study = c("TPD", "OPDC","PPMI"),
  effect = c("1.08", "1.01", "1.73"),
  ci = c("[0.57, 1.59]", "[0.23, 1.79]", "[0.71, 2.27]")
  )
summary = tibble(mean = 1.13,
                 lower = 1.14 - (1.96*0.2),
                 upper = 1.14 + (1.96*0.2),
                 study = "Summary",
                 effect = "1.14",
                 ci =  "[0.75 ,1.53]",
                 summary = TRUE)
header = tibble(study = "Study",
                effect = "β",
                ci = "CI",
                summary = TRUE)
empty_row <- tibble(mean = NA_real_)
SNPdf <- bind_rows(header,
                   SNP_cohort,
                   empty_row,
                   summary)



a = myforestplot(df = SNPdf, 
                 snpname = "rs180924818", 
                  n_studies = 3)

# Saving pdf
pdf("../FIGURES/rs180924818_forestPlot.pdf", width = 9)
a
dev.off()

# Saving jpg
jpeg("../FIGURES/rs180924818_forestPlot.jpg",
    width = 7.5, height = 6.5, units = "in", res = 500)
a 
dev.off()

```
