==> workflows/7. kmcurves.Rmd <==
---
title: 'Generate KM curves'
author: 
- name: "Alejandro M.Carrasco"
  affiliation: UCL
output: 
  html_document:
    code_folding: hide
    theme: paper
    highlight: kate
    df_print: paged
    toc: true
    toc_float: true
    number_sections: true
---

```{r setup, include=FALSE}
library(tidyverse)      # For tidy manipulation of data
library(here)           # For file path construction
library(DT)             # To display data tables
library(readxl)         # To read excel files
library(stringr)        # To do string operations
library(data.table)
library(survminer)
library(survival)

theme_rhr <-  theme_bw(base_family = "Helvetica") + 
  theme(panel.grid.major.x = element_blank(),
        legend.position = "right",
        strip.text = element_text(size = 8),
        strip.text.y = element_text(angle = 90),
        axis.text.x = element_text(size = 8, angle = 90, hjust = 1, vjust = 0.5),
        axis.text.y = element_text(size = 8),
        axis.title.y = element_text(vjust = 0.6),
        axis.title = element_text(size = 10),
        panel.spacing = unit(0.1, "lines"))
knitr::opts_chunk$set(echo = T, warning = F, message= F)
```


> Aim: To merge all data file from different cohorts and generate KM curves

# Data load

There is no need of doing any QC. Data has already been QCed.
We will do minor processing, and then merge
```{r}

# Load data
tpd = fread("../data/PROBAND_ALL.txt")
opdc = fread("../data/OPDC_ALL.txt")
ppmi = fread("../data/PPMI_ALL.txt")
pdbp = fread("../data/PDBP_ALL.txt")
pdstat = fread("../data/PDSTAT_ALL.txt")


```


# Merge the data 
```{r}
new_names = c("chr1_39646765", "chr1_53778300", "chr1_168645690",
"chr1_168645707", "chr4_32376657", "chr4_32435284", "chr16_17044975")
tpd_qc = tpd %>%
  dplyr::filter(left_censored == 0 ) %>%
  dplyr::select(IID, event_dyskinesia, time_Event = time_Event_midpoint,
                SEX, AAO = age_onset_imput, DISDUR = disease_duration_onset_imput, 
                MDSUPDRSIII = timeEvent_UPDRSIIItotal_imputed,
                LEDD = timeEvent_LdopaDose, starts_with("PC"),
                22:27) %>%
  rename_at(vars(colnames(tpd)[grepl(pattern = "_.$", colnames(tpd))]), ~new_names[2:length(new_names)])

# Rename strings



opdc_qc = opdc %>%
  dplyr::select(IID, event_dyskinesia, time_Event,
                SEX, AAO = age_onset, DISDUR = disease_duration_onset_BL, 
                MDSUPDRSIII = UPDRSIII_imputed_BL,
                LEDD = LEDD_BL, starts_with("PC"),
                19:25) %>%
  rename_at(vars(colnames(opdc)[grepl(pattern = "_.$", colnames(opdc))]), ~new_names)

ppmi_qc = ppmi %>%
  dplyr::select(IID = PATNO, event_dyskinesia = event_dysk, time_Event,
                SEX, AAO = ageonset, DISDUR = durOnset, 
                MDSUPDRSIII = updrs3_score_on, LEDD, starts_with("PC"),
                21:26) %>%
  dplyr::mutate(event_dyskinesia = as.numeric(event_dyskinesia)) %>%
  rename_at(vars(colnames(ppmi)[grepl(pattern = "_.$", colnames(ppmi))]), ~new_names[1:(length(new_names) - 1)])

pdbp_qc = pdbp %>%
  dplyr::filter(LONG_DISDUR == 0) %>% 
  dplyr::select(IID = ID, event_dyskinesia = EVENT_DYSK, 
                time_Event = TIMEMIDPOINT_YEARS, SEX = SEX.x, 
                AAO = AAD_IMPUTED, DISDUR = DIS_DUR, 
                MDSUPDRSIII = MDSUPDRSIII_BL, LEDD = DailyLevo,
                starts_with("PC"), 29:35) %>%
  rename_at(vars(colnames(pdbp)[grepl(pattern = "_.$", colnames(pdbp))]), ~new_names)

pdstat_qc = pdstat %>%
  dplyr::filter(longdisdur == 0) %>% 
  dplyr::select(IID, event_dyskinesia, time_Event,
                SEX, AAO, DISDUR = disDur, 
                MDSUPDRSIII, starts_with("PC"), 19:24) %>%
  rename_at(vars(colnames(pdstat)[grepl(pattern = "_.$", colnames(pdstat))]), ~new_names[c(1, 3:length(new_names))])


merge_nopdbp = dplyr::bind_rows(tpd_qc, opdc_qc, ppmi_qc, pdstat_qc)
merge_all = dplyr::bind_rows(tpd_qc, opdc_qc, ppmi_qc,
                         pdbp_qc, pdstat_qc)


#perform minor processing
merge_all = merge_all %>%
  mutate(AAO.std = scale(AAO),
         DISDUR.std = scale(DISDUR),
         MDSUPDRSIII.std = scale(MDSUPDRSIII),
         LEDD.std = scale(LEDD))
merge_nopdbp = merge_nopdbp %>%
  mutate(AAO.std = scale(AAO),
         DISDUR.std = scale(DISDUR),
         MDSUPDRSIII.std = scale(MDSUPDRSIII),
         LEDD.std = scale(LEDD))


summary(merge_nopdbp)
snps = colnames(merge_nopdbp)[14:20]
lapply(snps, function(mysnp) {
  merge_nopdbp %>% 
    group_by(.data[[mysnp]]) %>% 
    dplyr::summarise(count = n())
  })

```


# Get BMI and smoking status data available

```{r}
# In PPMI, smoking status is avialable. Number very slow (~40 individuals matching my LiD subset)
# Conclusion -> I will not include this data
ppmi_smoke = fread("~/Data/WorkingDirectory/PPMI/clinical_data/FOUND_RFQ_Smoking_History.csv")
ppmi_smoke$patno = paste0("PPMISI", ppmi_smoke$patno)
ppmi_smoke = ppmi_smoke %>%
  dplyr::select(PATNO = patno, smq2)
ppmi_smoke_dysk = ppmi %>%
  inner_join(ppmi_smoke, by = c("PATNO"))


#BMI data
ppmi_bmi = fread("~/Data/WorkingDirectory/PPMI/clinical_data/Vital_Signs.csv")
ppmi_bmi$PATNO = paste0("PPMISI", ppmi_bmi$PATNO)
ppmi_bmi = ppmi_bmi %>%
  dplyr::filter(EVENT_ID == "BL") %>%
  dplyr::select(IID = PATNO, WGTKG, HTCM)

ppmi_bmi_dysk = merge_all %>%
  inner_join(ppmi_bmi, by = c("IID")) %>%
  mutate(BMI = (WGTKG / HTCM / HTCM) * 10000,
         BMI_GROUP = ifelse(BMI < 18.5, "Low",
                            ifelse((BMI > 18.5) & (BMI < 24.9), "Healthy", 
                                   ifelse( (BMI > 25.0) & (BMI < 30), "Overweight",
                                          ifelse(BMI >= 30.0, "Obese", NA))))) %>%
  
  dplyr::select(IID, event_dyskinesia, time_Event, BMI_GROUP)



# OPDC 
# Not available

# TPD
# Get BMI and smoking status
tpd_clindemo_bl = fread("/home/amcalejandro/Data/WorkingDirectory/PROBAND/clinical_data/P3_full_extract_with_relations.csv")
get_smoking_bmi = tpd %>%
  dplyr::filter(left_censored == 0 ) %>%
  dplyr::select(IID, ID) %>%
  inner_join(tpd_clindemo_bl %>% 
               select(ID, BMI, merqp_smoking, smoke_QRISK))

tpd_smokingbmi = merge_all %>%
  inner_join(get_smoking_bmi) %>%
  mutate(BMI_GROUP = ifelse(BMI < 18.5, "Low",
                            ifelse((BMI > 18.5) & (BMI < 24.9), "Healthy", 
                                   ifelse( (BMI > 25.0) & (BMI < 30), "Overweight",
                                          ifelse(BMI >= 30.0, "Obese", NA)))),
          merqp_smoking = ifelse(merqp_smoking == 0, 1, merqp_smoking),
          SMOKING_GROUP = ifelse(merqp_smoking == 1, "Never or passive",
                            ifelse(merqp_smoking == 2, "Previous", 
                                   ifelse(merqp_smoking == 3, "Current",
                                          ifelse(BMI >= 30.0, "Obese", NA)))))



```


# Get hazard ratios on a model with confounding variables only
```{r}
survobject_all = Surv(time = merge_all$time_Event, event = merge_all$event_dyskinesia)
survobject_nopdbp = Surv(time = merge_nopdbp$time_Event, event = merge_nopdbp$event_dyskinesia)


# Fit best model
best_all_coxph = coxph(survobject_all ~ SEX + AAO + PC1 +
                          PC2 + PC3 + PC4 + PC5, data = merge_all)
summary(best_all_coxph)

basic_nopdbp_coxph = coxph(survobject_nopdbp ~ SEX + AAO + PC1 +
                                         PC2 + PC3 + PC4 + PC5, data = merge_nopdbp)
summary(basic_nopdbp_coxph)

```


# Generate surv objects and create basic models, then compare

```{r}
# Create survival object
survobject_all = Surv(time = merge_all$time_Event, event = merge_all$event_dyskinesia)
survobject_nopdbp = Surv(time = merge_nopdbp$time_Event, event = merge_nopdbp$event_dyskinesia)


## NOTE ##
# To ease covariates interpretation, we will use non standardised covariates to fit the models

# Fit basic cox PH model
basic_all_coxph = coxph(survobject_all ~ SEX + AAO + PC1 +
                                         PC2 + PC3 + PC4 + PC5, data = merge_all)
basic_nopdbp_coxph = coxph(survobject_nopdbp ~ SEX + AAO + PC1 +
                                         PC2 + PC3 + PC4 + PC5, data = merge_nopdbp)
#summary(basic_all_coxph)
#summary(basic_nopdbp_coxph)
# Check coxph assumption
#cox.zph(basic_all_coxph)
#cox.zph(basic_nopdbp_coxph)


# Fit best model
best_all_coxph = coxph(survobject_all ~ SEX + AAO + PC1 +
                          PC2 + PC3 + PC4 + PC5 +
                          DISDUR + MDSUPDRSIII, data = merge_all)

best_nopdbp_coxph = coxph(survobject_nopdbp ~ SEX + AAO + PC1 +
                             PC2 + PC3 + PC4 + PC5 +
                             DISDUR + MDSUPDRSIII, data = merge_nopdbp)
#summary(best_all_coxph)
#summary(best_nopdbp_coxph)

# Compare basic and best
AIC(basic_all_coxph, best_all_coxph)
AIC(basic_nopdbp_coxph, best_nopdbp_coxph)


# Fit best model plus LEDD
# Add other variables to the basic model
best_all_coxph = coxph(survobject_all ~ SEX + AAO + PC1 +
                          PC2 + PC3 + PC4 + PC5 +
                          DISDUR + MDSUPDRSIII + LEDD, data = merge_all)

best_nopdbp_coxph = coxph(survobject_nopdbp ~ SEX + AAO + PC1 +
                             PC2 + PC3 + PC4 + PC5 +
                             DISDUR + MDSUPDRSIII + LEDD, data = merge_nopdbp)
summary(best_all_coxph)
summary(best_nopdbp_coxph)

### Writing basic summary coefficient tables
write.csv(coefficients(summary(best_nopdbp_coxph)), 
            file = "../data_processed/ALL_coeeficients_bestModel_nopdbp.csv")
### Writing basic summary coefficient tables
write.csv(coefficients(summary(basic_nopdbp_coxph)), 
            file = "../data_processed/ALL_coeeficients_basicModel_nopdbp.csv")
```


# Add SNPs to the previous survival objects 
```{r}
# Create a model for each snp and generate KM curves
   # Get vector of SNP names
vector_snps = colnames(merge_all)[startsWith(colnames(merge_all), "chr")]
fit_basic = formula("survobject_nopdbp ~ SEX + AAO.std + 
                                     PC1 + PC2 + PC3 + 
                                     PC4 + PC5")
fit_best = formula("survobject_nopdbp ~ SEX + AAO.std + PC1 +
                                    PC2 + PC3 + PC4 + PC5 +
                                    DISDUR.std + MDSUPDRSIII.std")


basic_object_all= coxph(fit_basic, data = merge_nopdbp)
best_object_all = coxph(fit_best, data = merge_nopdbp)


new_list = list()
for (p in vector_snps) {
  # Update formula with the Pth SNP
  formula = stats::as.formula(paste(". ~ . + ", p))

  basic_update = stats::update(basic_object_all, formula)
  best_update = stats::update(best_object_all, formula)
  
  merge = dplyr::bind_rows(coefficients(summary(basic_update))[,5], 
                coefficients(summary(best_update))[,5])
  
  merge = cbind(merge, MODEL= c("model_basic", "model_best"))
  new_list[[p]] = merge
  
}

saveRDS(new_list, "../COVARIATES_Pvalues_MODELS.rds")
```


# Save KM curves for the different SNPs

```{r}
merge_nopdbp = merge_nopdbp %>%
  mutate(chr1_53778300 = ifelse(chr1_53778300 == 2, 1, chr1_53778300),
         chr4_32435284 = ifelse(chr4_32435284 == 2, 1, chr4_32435284),
         chr16_17044975 = ifelse(chr16_17044975 == 2, 1, chr16_17044975)) %>%
  dplyr::rename(rs72673189=chr1_53778300,
                rs189093213=chr4_32435284,
                rs180924818=chr16_17044975)


myvars = colnames(merge_nopdbp)[c(14,18,19)]
lapply(myvars, FUN = function(x) {
  myformula = formula(paste("survobject_nopdbp ~ ", x ))
  fit_snps_dyskinesia <- do.call(survfit, args = list(formula = myformula, data = merge_nopdbp))
  #fit <- do.call(survfit, args = list(formula = formula, data = lung))
  #jpeg()
  #myplot = ggsurvplot(fit_snps_dyskinesia, data = merge_nopdbp, pval = TRUE, xlab = "years", ylab = "Survival probability")
  
  #parse p-value 
  pval <- surv_pvalue(fit = fit_snps_dyskinesia , data = merge_nopdbp)


  myplot =  ggsurvplot(fit_snps_dyskinesia, data = merge_nopdbp,
                       #pval = TRUE,
                       pval = formatC(pval$pval, format = "e", digits = 3),
                       pval.size = 6,
                       legend.title = x, 
                       risk.table = TRUE, tables.height = 0.2,
                       conf.int = TRUE,
                       tables.theme = theme_cleantable(),
                       palette = c("#E7B800", "#2E9FDF"),
            ggtheme = theme_bw(),
            xlim = c(0,17),
            legend = "bottom",
            font.legend = 10,
            font.x = c(14, "bold"),
            font.tickslab = 14,
            font.y = c(14, "bold"),
            font.title = c(20, "bold"),
            xlab = "years", title = paste(x, "KM curves"))
  

  
  jpeg(paste0("../FIGURES/plot_", x ,".jpg"), width = 7.5, height = 6.5, units = 'in', res = 500)
  print(myplot)
  dev.off()

  
  tiff(filename = paste0("../FIGURES/plot_", x ,".tiff"), units="in", width=5, height=5, res=300)
  print(myplot)
  dev.off()
  
  pdf(file = paste0("../FIGURES/plot_", x ,".pdf"), width = 12, height = 8)
  print(myplot)
  dev.off()
  
} )
```



# Test the effect of gender in the merged data

```{r}

# Create survival object
survobject_all = Surv(time = merge_nopdbp$time_Event, event = merge_nopdbp$event_dyskinesia)
basic_all_gender = survfit(survobject_all ~ SEX, data = merge_nopdbp)
coxph_hr_aao = coxph(survobject_nopdbp ~ SEX, data = merge_nopdbp)

#parse p-value 
pval <- surv_pvalue(fit = basic_all_gender, data = merge_nopdbp)

myplot = ggsurvplot(basic_all_gender, data = merge_nopdbp,
            pval = round(pval$pval, digits = 3),
            legend.title = "Sex", legend.labs = c("Male", "Female"),
            risk.table = TRUE, tables.height = 0.2,
            conf.int = TRUE,
            tables.theme = theme_cleantable(),
            palette = c("#E7B800", "#2E9FDF"),
            ggtheme = theme_bw(),
            xlim = c(0,17),
            legend = "bottom",
            font.legend = 10,
             font.x = c(14, "bold"),
            font.tickslab = 14,
            font.y = c(14, "bold"),
            font.title = c(20, "bold"),
            xlab = "years", title = "Gender KM curves")

jpeg(paste0("../FIGURES/plot_SURVIVAL_SEX.jpg"), width = 7.5, height = 6.5, units = 'in', res = 500)
print(myplot)
dev.off()

tiff(filename = paste0("../FIGURES/plot_SURVIVAL_SEX.tiff"), units="in", width=5, height=5, res=300)
print(myplot)
dev.off()
  

# Extract the hazard ratio
coxph_hr_aao =
  coxph(survobject_nopdbp ~ SEX + AAO + PC1 +
          PC2 + PC3 + PC4 + PC5, data = merge_nopdbp)
#summary(basic_all_coxph)

```


# Test the effect of AAO in the merged data

```{r}
# Get strat AAO
merge_nopdbp = merge_nopdbp %>% 
  mutate(AAO_GROUP = ifelse(AAO > 50, "Old", "Young"))

# Create survival object
survobject_nopdbp = Surv(time = merge_nopdbp$time_Event, event = merge_nopdbp$event_dyskinesia)
basic_nopdbp_AAO = survfit(survobject_nopdbp ~ AAO_GROUP, data = merge_nopdbp)

pval <- surv_pvalue(fit = basic_nopdbp_AAO, data = merge_nopdbp)

            
myplot = ggsurvplot(basic_nopdbp_AAO, data = merge_nopdbp,
            #pval = formatC(pval$pval, format = "e", digits = 3),
            pval = round(pval$pval, digits = 5),
            legend.title = "AAO", legend.labs = c("AAO > 50", "AAO ≤ 50"),
            risk.table = TRUE, tables.height = 0.2,
            conf.int = TRUE,
            tables.theme = theme_cleantable(),
            palette = c("#E7B800", "#2E9FDF"),
            ggtheme = theme_bw(),
            xlim = c(0,17),
            legend = "bottom",
            font.legend = 10,
             font.x = c(14, "bold"),
            font.tickslab = 14,
            font.y = c(14, "bold"),
            font.title = c(20, "bold"),
            xlab = "years", title = "AAO groups KM curves")

jpeg(paste0("../FIGURES/plot_SURVIVAL_AAO_GROUP_NOPDBP.jpg"), width = 7.5, height = 6.5, units = 'in', res = 500)
print(myplot)
dev.off()

tiff(filename = paste0("../FIGURES/plot_SURVIVAL_AAO_GROUP_NOPDBP.tiff"), units="in", width=5, height=5, res=300)
print(myplot)
dev.off()
```

# Test BMI and smoking status in TPD

```{r}

# Get KM curve for smoking status from TPD
survobject_nopdbp = Surv(time = tpd_smokingbmi$time_Event, event = tpd_smokingbmi$event_dyskinesia)
basic_tpd_smoking = survfit(survobject_nopdbp ~ SMOKING_GROUP, data = tpd_smokingbmi)
myplot = ggsurvplot(basic_tpd_smoking, data = tpd_smokingbmi,
            pval = TRUE,
            legend.title = "SMOKING", #legend.labs = c("Healthy", "Underweight", "Obese", "Overweight"),
            risk.table = TRUE, tables.height = 0.2,
            conf.int = TRUE,
            tables.theme = theme_cleantable(),
            palette = c("#E7B800", "#2E9FDF", "#B2182B"),
            ggtheme = theme_bw(),
            xlim = c(0,17),
            legend = "bottom",
            font.legend = 10,
             font.x = c(14, "bold"),
            font.tickslab = 14,
            font.y = c(14, "bold"),
            font.title = c(20, "bold"),
            xlab = "years", title = "SMOKING groups KM curves")
jpeg(paste0("../FIGURES/plot_SURVIVAL_SMOKING_GROUP_TPD.jpg"), width = 7.5, height = 6.5, units = 'in', res = 500)
print(myplot)
dev.off()
tiff(filename = paste0("../FIGURES/plot_SURVIVAL_SMOPKING_GROUP_TPD.tiff"), units="in", width=5, height=5, res=300)
print(myplot)
dev.off()


# Create survival object
# Merge TP and PPMI BMI data
bmi_all = rbind(ppmi_bmi_dysk, tpd_smokingbmi %>% dplyr::select(IID, event_dyskinesia, time_Event, BMI_GROUP))

survobject_nopdbp = Surv(time = bmi_all$time_Event, event = bmi_all$event_dyskinesia)
basic_tpd_bmi = survfit(survobject_nopdbp ~ BMI_GROUP, data = bmi_all)
myplot = ggsurvplot(basic_tpd_bmi, data = bmi_all,
            pval = TRUE,
            legend.title = "BMI", legend.labs = c("Healthy", "Underweight", "Obese", "Overweight"),
            risk.table = TRUE, tables.height = 0.2,
            conf.int = TRUE,
            tables.theme = theme_cleantable(),
            palette = c("#E7B800", "#2E9FDF", "#B2182B", "#2166AC"),
            ggtheme = theme_bw(),
            xlim = c(0,17),
            legend = "bottom",
            font.legend = 10,
             font.x = c(14, "bold"),
            font.tickslab = 14,
            font.y = c(14, "bold"),
            font.title = c(20, "bold"),
            xlab = "years", title = "BMI groups KM curves")

jpeg(paste0("../FIGURES/plot_SURVIVAL_BMI_GROUP_TPD.jpg"), width = 7.5, height = 6.5, units = 'in', res = 500)
print(myplot)
dev.off()
tiff(filename = paste0("../FIGURES/plot_SURVIVAL_BMI_GROUP_TPD.tiff"), units="in", width=5, height=5, res=300)
print(myplot)
dev.off()


```
