==> workflows/8. sensitivityAnalysis.Rmd <==
---
title: 'Sensitivity analysis'
author: 
- name: "Alejandro M.Carrasco"
  affiliation: UCL
output: 
  html_document:
    code_folding: hide
    theme: paper
    highlight: kate
    df_print: paged
    toc: true
    toc_float: true
    number_sections: true
---


# Load data and pkgs
```{r}
library(dplyr)
library(data.table)
library(tidyverse)
library(survminer)
library(survival)
```


# Load Tracking Parkinson (TP) data
```{r}
genotypes = fread("../data/sensitivty_analysis/PROBAND_LID_HITS_SENSITIVITY.raw") %>%
  dplyr::select(c(2,7:9))
clinical = readRDS("../data/sensitivty_analysis/DyskTimToEvent_UPDRS_3.cov_motorSEX.RDS")
PCs <- read.table("../data/sensitivty_analysis/PCA.eigenvec", sep = "", header= FALSE )%>%
  dplyr::select(V2:V7) %>%
  dplyr::rename(IID = V2,
         PC1 = V3,
         PC2 = V4,
         PC3 = V5,
         PC4 = V6,
         PC5 = V7)
```


# Do minor processing

```{r}
tpd= clinical %>%
  inner_join(PCs) %>%
    inner_join(genotypes)

new_names = c("chr1_53778300", "chr4_32435284", "chr16_17044975")
tpd_qc = tpd %>%
  filter(!left_censored == 1) %>%
  dplyr::select(IID, event_dyskinesia, 
                time_Event = time_Event_midpoint, SEX,
                AAO = age_onset_imput,
                DISDUR = disease_duration_onset_imput, 
                MDSUPDRSIII = timeEvent_UPDRSIIItotal_imputed,
                ldopa = timeEvent_LdopaDose, starts_with("PC"),
                22:24) %>%
  rename_at(vars(colnames(tpd)[grepl(pattern = "_.$", colnames(tpd))]), ~new_names)
```


# Explore the significance of SNPs in tpd with and without ldopa adjustment at the time of event

```{r}
survobject = Surv(time = tpd_qc$time_Event, event = tpd_qc$event_dyskinesia)

vector_snps = colnames(tpd_qc)[startsWith(colnames(tpd_qc), "chr")]
fit_noldopa = formula("survobject ~ SEX + AAO + 
                                     PC1 + PC2 + PC3 + 
                                     PC4 + PC5")
fit_ldopa = formula("survobject ~ SEX + AAO + PC1 +
                                    PC2 + PC3 + PC4 + PC5 +
                                    ldopa")

noldopa_object = coxph(fit_noldopa, data = tpd_qc)
ldopa_object = coxph(fit_ldopa, data = tpd_qc)


new_list = list()
for (p in vector_snps) {
  # Update formula with the Pth SNP
  formula = stats::as.formula(paste(". ~ . + ", p))

  noldopa_update = stats::update(noldopa_object, formula)
  ldopa_update = stats::update(ldopa_object, formula)
  
  ldopa_coef = coefficients(summary(ldopa_update))[,5]
  
  merge = dplyr::bind_rows(coefficients(summary(noldopa_update))[,5], 
                           coefficients(summary(ldopa_update))[,5]) %>%
    dplyr::mutate(SNP = p) %>%
    dplyr::rename(SNP_effect = .data[[p]])
    
  merge = cbind(merge, MODEL= c("noldopa_model", "ldopa_model"))
  new_list[[p]] = merge
}


final_df = map_dfr(new_list, bind_rows)
write_csv(final_df, "../sensitivity_out/sensitivity_ldopa_tpd.csv")
```


# Explore the significance of SNPs in tpd with and without motor adjustement when the time of event

```{r}
survobject = Surv(time = tpd_qc$time_Event, event = tpd_qc$event_dyskinesia)
vector_snps = colnames(tpd_qc)[startsWith(colnames(tpd_qc), "chr")]
fit_nomotor = formula("survobject ~ SEX + AAO + 
                                     PC1 + PC2 + PC3 + 
                                     PC4 + PC5")
fit_motor = formula("survobject ~ SEX + AAO + PC1 +
                                    PC2 + PC3 + PC4 + PC5 +
                                    MDSUPDRSIII")

nomotor_object = coxph(fit_nomotor, data = tpd_qc)
motor_object = coxph(fit_motor, data = tpd_qc)


new_list = list()
for (p in vector_snps) {
  # Update formula with the Pth SNP
  formula = stats::as.formula(paste(". ~ . + ", p))

  nolmotor_update = stats::update(nomotor_object, formula)
  motor_update = stats::update(motor_object, formula)
  
  motor_coef = coefficients(summary(motor_update))[,5]
  
  merge = dplyr::bind_rows(coefficients(summary(nomotor_update))[,5], 
                           coefficients(summary(motor_update))[,5]) %>%
    dplyr::mutate(SNP = p) %>%
    dplyr::rename(SNP_effect = .data[[p]])
    
  merge = cbind(merge, MODEL= c("nomotor_model", "motor_model"))
  new_list[[p]] = merge
}

final_df = map_dfr(new_list, bind_rows)
write_csv(final_df, "../sensitivity_out/sensitivity_motor_tpd.csv")
```


# Explore the meta-analysis hits excluding PDBP data

```{bash}
grep "NOPDBP.txt" ~/RES/METAANALYSIS/ |  
  awk '{print $2}' | 
  xargs grep -E "4:32435284|16:17044975|1:53778300" >> snps_metrics_nopdbp.txt
```
