==> workflows/14. stepwiseSelection.Rmd <==
---
title: 'Develop Stepwise Selection'
author: 
- name: "Alejandro M.Carrasco"
  affiliation: UCL
output: 
  html_document:
    code_folding: hide
    theme: paper
    highlight: kate
    df_print: paged
    toc: true
    toc_float: true
    number_sections: true
---



# Set up

```{r}
.libPaths("/mnt/rreal/RDS/acarrasco/R_libs")
library(tidyverse)      # For tidy manipulation of data
library(here)           # For file path construction
library(DT)             # To display data tables
library(readxl)         # To read excel files
library(stringr)        # To do string operations
library(data.table)
library(survminer)
library(survival)
```


# Load data
```{r}
# Load data
tpd = fread("../data/PROBAND_ALL.txt")
tpd_clindemo_bl = fread("/home/amcalejandro/Data/WorkingDirectory/PROBAND/clinical_data/P3_full_extract_with_relations.csv")
tpd_clindemo_lng = fread("/home/amcalejandro/Data/WorkingDirectory/PROBAND/clinical_data/P3_TPD_Version2_temp_06_16_2020.csv")
tpd_ldopabin = fread("/home/amcalejandro/Data/WorkingDirectory/PROBAND/clinical_data/Tracking_Ldopa_data_2021_03.csv")
tpd_ldopadoses = fread("/home/amcalejandro/Data/WorkingDirectory/PROBAND/clinical_data/Tracking_Ldopa_data_with_doses_2021_03.csv")
#tpd_nfl = fread("/home/amcalejandro/Data/WorkingDirectory/PROBAND/clinical_data/Proband_NfL.txt")
tpd_meds = readxl::read_xlsx("/home/amcalejandro/Data/WorkingDirectory/PROBAND/clinical_data/P3_LEDD_all_visits_PD_proba_V1.xlsx")
```



# Get all baseline data and drop variables with high missingness rate

QC and merge
```{r}

new_names = c("chr1_39646765", "chr1_53778300", "chr1_168645690",
"chr1_168645707", "chr4_32376657", "chr4_32435284", "chr16_17044975")
tpd_qc = tpd %>%
  dplyr::filter(left_censored == 0) %>%
  dplyr::select(ID, IID, event_dyskinesia, time_Event = time_Event_midpoint,
                SEX, AAO = age_onset_imput, 
                MDSUPDRSIII = timeEvent_UPDRSIIItotal_imputed, starts_with("PC"),
                #LEDD = timeEvent_LdopaDose,  DISDUR = disease_duration_onset_imput 
                22:27) %>%
  mutate(AAO.std = scale(AAO))%>%
  rename_at(vars(colnames(tpd)[grepl(pattern = "_.$", colnames(tpd))]), ~new_names[2:length(new_names)])

tpd_ldopabin_bl = tpd_ldopabin %>%
  filter(Visit == 1) %>%
  select(-c(Visit, Ldopa_bin))

tpd_ldopadoses_bl = tpd_ldopadoses %>%
  filter(Visit == 1) %>%
  select(-c(Visit, age, )) %>%
  mutate(Ldopa_bin = as.factor(Ldopa_bin))


tpd_meds_bl = tpd_meds %>%
  select(ID, ends_with(c("PD_meds", "V1")), -Ldopa_bin_V1)


tpd_clindemo_bl = tpd_clindemo_bl %>%
  select(-c(gender, diag_date, onset_date, visit_date, age, age_diag,
            age_onset, disease_duration_diag))

tpd_all = tpd_qc %>%
  inner_join(tpd_ldopabin_bl, by = "ID") %>%
  inner_join(tpd_ldopadoses_bl, by = "ID") %>%
  inner_join(tpd_meds_bl, by = "ID") %>%
  inner_join(tpd_clindemo_bl %>%
               select(-c(Ldopa_bin, Ldopa_previous)), by = "ID")
```



# Explore missingness

If one variable has more than 10% missingness, I drop it

```{r}
missingness_50 = dim(tpd_all)[1] * 0.10
get_missingness = setNames(
  sapply(tpd_all, function(x) sum(is.na(x))) %>%
    as.data.frame() %>% rownames_to_column(),
  c("VARS", "MISSINGNESS")) %>%
  mutate(to_exclude = ifelse(MISSINGNESS > missingness_50, 1, 0))


# Get the names of those variables with a very high missigness
high_missingness = get_missingness %>%
  filter(to_exclude == 1)

valid_cols = get_missingness %>%
  filter(to_exclude ==0) %>%
  pull(VARS)
tpd_all = tpd_all %>%
  select(all_of(valid_cols))

# Load PRS and merge with TPD data
tpd_prs =  tpd_prs %>%
  dplyr::rename(prs_sginificant = `Pt_5e-08`, prs_nominal = `Pt_5e-06`)
tpd_all = tpd_all %>%
  inner_join(tpd_prs %>% dplyr::select(c(IID, prs_nominal)))

cat("Number of variables with low missingness rate before trimming", dim(tpd_all)[2])
cat("\n")


```


# Trim list of covariates and and creater base model
```{r}
# trim the list of BL covariates
tpd_all = tpd_all %>%
  dplyr::select(!((starts_with("PDQ8")) &  !(ends_with("total")) |
                    (starts_with("ESS")) &  !(ends_with("total")) |
                    (starts_with("RBD")) &  !(ends_with("total")) |
                    (starts_with("QUIP")) &  !(ends_with("all")) |
                    (starts_with("MOCA")) &  !(ends_with("total")) |
                    (starts_with("RBD")) &  !(ends_with("total")) |
                    (starts_with("PDSS")) &  !(ends_with("total")) |
                    (starts_with("UPDRS_I")) &  !(ends_with("total")) |
                    (starts_with("UPDRS_II")) &  !(ends_with("total")) |
                    (starts_with("UPDRS_III")) |
                    (starts_with("UPDRS_")) &  !(contains(c("I", "II", "III", "IV"))))) %>%
  dplyr::select(!(c(starts_with(c("present_", "v3_", "v7", "UPDRS_III")),
                    "no_neuroimaging", "Withdrawn", "change_diagnosis", "Related_study_IDs")))

cat("Number of variables with low missingness rate after trimming", dim(tpd_all)[2])
cat("\n")


fit_formula = formula("event_dyskinesia ~ AAO.std + SEX +
                      PC1 + PC2 + PC3 + PC4 + PC5")
tpd_all = tpd_all %>%
  select(-c(IID, time_Event, AAO))

vector_covariates = colnames(tpd_all)[!colnames(tpd_all)
                                      %in% c("event_dyskinesia", "ID")]
count_nonas = function(x) {
  unique(na.omit(x))
}

get_levels = sapply(lapply(tpd_all %>% dplyr::select(vector_covariates), count_nonas), length)
novariance_vars = names(which(get_levels == 1)) # Check there are no variables with only one level
vector_covariates = vector_covariates[!vector_covariates %in% novariance_vars]

```


# Run stepwise regression algorithm 

```{r}
source(here::here("utils", "utils.R")) # load the custom stepwise reg algorithm

fitlogmodel = glm(fit_formula, data = tpd_all)
upper = vector_covariates
fit = fitlogmodel

save = stepper(fit = fitlogmodel, upper = vector_covariates)
saveRDS(save, "results_stepWiseReg_less10missing_GLM_PRS.rds")
```
