==> workflows/15.2. anxiety_tpd_ppmi.Rmd <==
---
title: 'Explore anxiety metrics in TPD and PPMI'
author: 
- name: "Alejandro M.Carrasco"
  affiliation: UCL
output: 
  html_document:
    code_folding: hide
    theme: paper
    highlight: kate
    df_print: paged
    toc: true
    toc_float: true
    number_sections: true
---

# Load libraries
```{r setup, include=FALSE}
library(tidyverse)      # For tidy manipulation of data
library(here)           # For file path construction
library(DT)             # To display data tables
library(readxl)         # To read excel files
library(stringr)        # To do string operations
library(data.table)
library(survminer)
library(survival)

```

# Data load
```{r}

# Load data
tpd = fread("../data/PROBAND_ALL.txt")
opdc = fread("../data/OPDC_ALL.txt")
ppmi = fread("../data/PPMI_ALL.txt")
pdbp = fread("../data/PDBP_ALL.txt")
pdstat = fread("../data/PDSTAT_ALL.txt")


```


# Process and merge all the data

```{r}
new_names = c("chr1_39646765", "chr1_53778300", "chr1_168645690",
"chr1_168645707", "chr4_32376657", "chr4_32435284", "chr16_17044975")
tpd_qc = tpd %>%
  dplyr::filter(left_censored == 0 ) %>%
  dplyr::select(IID, event_dyskinesia, time_Event = time_Event_midpoint,
                SEX, AAO = age_onset_imput, DISDUR = disease_duration_onset_imput, 
                MDSUPDRSIII = timeEvent_UPDRSIIItotal_imputed,
                LEDD = timeEvent_LdopaDose, starts_with("PC"),
                22:27) %>%
  rename_at(vars(colnames(tpd)[grepl(pattern = "_.$", colnames(tpd))]), ~new_names[2:length(new_names)])


ppmi_qc = ppmi %>%
  dplyr::select(IID = PATNO, event_dyskinesia = event_dysk, time_Event,
                SEX, AAO = ageonset, DISDUR = durOnset, 
                MDSUPDRSIII = updrs3_score_on, LEDD, starts_with("PC"),
                21:26) %>%
  dplyr::mutate(event_dyskinesia = as.numeric(event_dyskinesia)) %>%
  rename_at(vars(colnames(ppmi)[grepl(pattern = "_.$", colnames(ppmi))]), ~new_names[1:(length(new_names) - 1)])

```


# Get anxiety data from TPD and PPMI

```{r}
tpd_bl = fread("~/Data/PROBAND/clinical_data/P3_full_extract_with_relations.csv")
tpd_bl_anxiety = tpd_bl %>% 
  dplyr::select(c(ID, grep("anx", colnames(tpd_bl))))

tpd_qc = tpd %>%
  dplyr::filter(left_censored == 0 ) %>%
  dplyr::select(ID, ID, event_dyskinesia, time_Event = time_Event_midpoint,
                SEX, AAO = age_onset_imput, DISDUR = disease_duration_onset_imput, 
                MDSUPDRSIII = timeEvent_UPDRSIIItotal_imputed,
                LEDD = timeEvent_LdopaDose, starts_with("PC"),
                22:27) %>%
  rename_at(vars(colnames(tpd)[grepl(pattern = "_.$", colnames(tpd))]), ~new_names[2:length(new_names)])


tpd_bl_anxiety_qc = tpd_qc %>% inner_join(tpd_bl_anxiety)


ppmi_anxiety = fread("~/Data/PMI/clinical_data/State-Trait_Anxiety_Inventory.csv")
# Now, we are going to get the Scale total score (items 1-40) the scale state score (items 1-20) and the scale trait score (items 21-40)
ppmi_anxiety_qc = ppmi_anxiety %>%
  dplyr::filter(EVENT_ID == "BL") %>%
  dplyr::mutate(PATNO = paste0("PPMISI", PATNO)) %>%
  dplyr::filter(PATNO %in% unique(ppmi$PATNO)) %>%
  rowwise() %>% 
  mutate(state_score = sum(c_across(STAIAD1:STAIAD20)),
         trait_score = sum(c_across(STAIAD21:STAIAD40)),
         total_score = sum(c_across(STAIAD1:STAIAD40))) %>%
  dplyr::select(PATNO, EVENT_ID, ends_with("score"))
ppmi_anxiety_qc = ppmi_qc %>% inner_join(ppmi_anxiety_qc, by = c("IID" = "PATNO"))

```


# Create logistic model adding confounders and anxiety measures only
```{r}
# Get logistic model
anxiety_glm_tpd = glm(event_dyskinesia ~ SEX + AAO + PC1 +
                          PC2 + PC3 + PC4 + PC5 + Leeds_anx_bin, 
                      data = tpd_bl_anxiety_qc,
                      family = "binomial")

summary(anxiety_glm_tpd)

anxiety_glm_ppmi = glm(event_dyskinesia ~ SEX + AAO + PC1 +
                          PC2 + PC3 + PC4 + PC5 + trait_score, 
                      data = ppmi_anxiety_qc,
                      family = "binomial")

summary(anxiety_glm_ppmi)

anxiety_glm_ppmi_total = glm(event_dyskinesia ~ SEX + AAO + PC1 +
                          PC2 + PC3 + PC4 + PC5 + total_score, 
                      data = ppmi_anxiety_qc,
                      family = "binomial")

summary(anxiety_glm_ppmi_total)

```


# Conclussion 

Anxiety was not associated with LiD in PPMI data.
Anxiety is measured based on a different clinical assesment in PPMI, compared to Tracking data.
In addition, the sample size in PPMI is smaller, so it could be a matter of power as well differences in the underlying subset of the general PD population represented in PPMI and Tracking Parkincons cohorts
